Create Client Class, Reading Methods, and Add a Worksheet Selection Form
In this commit, the client class, its methods, and a new form will be added to allow the user to choose which worksheet to import from an Excel file. Depending on the user's selection, the application will read and display data from either the Debitos or Clientes worksheet, fulfilling the first part of the challenge, which is to enable the reading of both worksheets based on user choice.